Move withConfigurationInputTrackingDisabled to new UnsafeConfigurationService
Since it doesn't really have anything to do with background job executors.
