+++ platforms/enterprise/enterprise/src/main/java/org/gradle/internal/enterprise/GradleEnterprisePluginRequiredServices.java
@@ -32,4 +32,6 @@ public interface GradleEnterprisePluginRequiredServices {
     StyledTextOutputFactory getStyledTextOutputFactory();
 
     GradleEnterprisePluginBackgroundJobExecutors getBackgroundJobExecutors();
+
+    DevelocityPluginUnsafeConfigurationService getUnsafeConfigurationService();
 }
